requirements.txt
e3nn==0.5.6
lpips==0.1.4
matplotlib==3.9.2
nibabel==5.2.1
numpy==2.1.1
pytorch-lightning==2.4.0
PyYAML==6.0.1
scipy==1.14.1
setuptools==72.1.0
torch==2.4.1
torchmetrics==1.4.1
torchvision==0.19.1
tensorboardX==2.6.2.2
scikit-learn==1.6.1
pandas==2.2.3
dipy==1.11.0 

